move compare utils and some cleaners
