+++ src/components/GraphWatcher/Menus/DNPMenu.stories.tsx
@@ -1,10 +1,10 @@
 import { ComponentProps } from "react";
+import { expect } from "@storybook/jest";
 import { Meta, StoryObj } from "@storybook/react";
 import { userEvent, waitFor, within } from "@storybook/testing-library";
 
-import { expect } from "@storybook/jest";
 import DNPMenu from "./DNPMenu";
-import { cleanRelatedItem } from "./utils";
+import { cleanRelatedItem } from "../helpers";
 
 import { items } from "Mocks";
