+++ src/components/GraphWatcher/WebImport/helpers.test.tsx
@@ -1,9 +1,9 @@
 import { act, render } from "@testing-library/react";
 
-import { webimportClass } from "../classes";
-import { hasNodeListChanged } from "../../../utils";
-import { findWebimportDivs, setWebimportDivs } from "./utils";
+import { hasNodeListChanged } from "../helpers";
+import { findWebimportDivs, matchArrays, setWebimportDivs } from "./helpers";
 
+import { webimportClass } from "../classes";
 import { blockWithAliasedLink } from "../fixtures";
 
 
@@ -49,4 +49,21 @@ describe("WebImport divs are inserted", () => {
 			}
 		}
 	);
+});
+
+describe("matchArrays", () => {
+	it("Finds if two string arrays have elements in common", () => {
+		const arr1 = ["tools", "platforms", "models"];
+		const arr2 = ["makers", "founders", "tools"];
+		const arr3 = ["makers", "founders", "companies"];
+	
+		expect([
+			matchArrays(arr1, arr2),
+			matchArrays(arr1, arr3)
+		])
+			.toEqual([
+				true,
+				false
+			]);
+	});
 });
