+++ src/components/SearchPanel/helpers.ts
@@ -1,5 +1,23 @@
+import { categorizeLibraryItems, cleanLibraryItem, identifyChildren } from "../../utils";
+
 import { SettingsCopy } from "Types/extension";
-import { ZCleanItemTop } from "Types/transforms";
+import { RCitekeyPages, ZCleanItemTop, ZItem, ZItemTop } from "Types/transforms";
+
+
+/** Categorizes library data into top-level items, annotations/notes, and PDFs with their metadata, children, and links */
+function cleanLibrary(arr: ZItem[], roamCitekeys: RCitekeyPages): ZCleanItemTop[] {
+	const lib = categorizeLibraryItems(arr);
+
+	return lib.items
+		.map((item: ZItemTop) => {
+			const itemKey = item.data.key;
+			const location = item.library.type + "s/" + item.library.id;
+			const { pdfs, notes } = identifyChildren(itemKey, location, { pdfs: lib.pdfs, notes: lib.notes });
+
+			return cleanLibraryItem(item, pdfs, notes, roamCitekeys);
+		});
+}
+
 
 /** Creates a reference to a simplified item with default format */
 function formatItemReferenceWithDefault(simplifiedItem: ZCleanItemTop, copySettings: SettingsCopy): string {
@@ -54,6 +72,7 @@ function formatItemReferenceForCopy(
 }
 
 export {
+	cleanLibrary,
 	formatItemReferenceWithDefault
 };
