+++ src/components/ZoteroImport/helpers.test.ts
@@ -1,53 +1,6 @@
-import { 
-	compareItemsByYear,
-	matchArrays,
-	sortCollections,
-	sortElems } from "../../src/utils";
-
+import { sortCollections } from "./helpers";
 import { ZoteroAPI } from "Types/externals";
 
-
-test("Sorts Zotero items by publication year & creators", () => {
-	const a = { meta: { creatorSummary: "Smith et al.", parsedDate: "" } };
-	const b = { meta: { creatorSummary: "Scott et al.", parsedDate: "" } };
-	const c = { meta: { creatorSummary: "Smith et al.", parsedDate: "2022-01-01" } };
-	const d = { meta: { creatorSummary: "Tikki and Noald", parsedDate: "2021-01-01" } };
-	const e = { meta: { creatorSummary: "Chen and Talmanes", parsedDate: "2021" } };
-
-	expect([
-		[a, b].sort(compareItemsByYear),
-		[a, c].sort(compareItemsByYear),
-		[c, a].sort(compareItemsByYear),
-		[b, c].sort(compareItemsByYear),
-		[c, d].sort(compareItemsByYear),
-		[d, c].sort(compareItemsByYear),
-		[d, e].sort(compareItemsByYear)
-	])
-		.toEqual([
-			[b, a],
-			[c, a],
-			[c, a],
-			[c, b],
-			[d, c],
-			[d, c],
-			[e, d]
-		]);
-});
-
-test("Finds if two string arrays have elements in common", () => {
-	const arr1 = ["tools", "platforms", "models"];
-	const arr2 = ["makers", "founders", "tools"];
-	const arr3 = ["makers", "founders", "companies"];
-
-	expect([
-		matchArrays(arr1, arr2),
-		matchArrays(arr1, arr3)
-	])
-		.toEqual([
-			true,
-			false
-		]);
-});
 
 describe("Zotero collections sorting", () => {
 	it("returns an empty array when given an empty input", () => {
@@ -90,7 +43,7 @@ describe("Zotero collections sorting", () => {
 					"name": "LoC",
 					"parentCollection": false,
 					"relations": {
-		
+
 					}
 				}
 			},
@@ -132,7 +85,7 @@ describe("Zotero collections sorting", () => {
 					"name": "sherlock films",
 					"parentCollection": "QM6T3KHX",
 					"relations": {
-		
+
 					}
 				}
 			},
@@ -174,7 +127,7 @@ describe("Zotero collections sorting", () => {
 					"name": "Digital Newspaper project",
 					"parentCollection": "N7W92H48",
 					"relations": {
-		
+
 					}
 				}
 			},
@@ -212,7 +165,7 @@ describe("Zotero collections sorting", () => {
 					"name": "Non-English items",
 					"parentCollection": false,
 					"relations": {
-		
+
 					}
 				}
 			},
@@ -250,12 +203,12 @@ describe("Zotero collections sorting", () => {
 					"name": "pubmed",
 					"parentCollection": false,
 					"relations": {
-		
+
 					}
 				}
 			}
 		];
-	
+
 		expect(sortCollections(colls))
 			.toMatchObject([
 				{ key: "N7W92H48", depth: 0 },
@@ -265,24 +218,4 @@ describe("Zotero collections sorting", () => {
 				{ key: "2GUIGKC9", depth: 0 }
 			]);
 	});
-});
-
-test("Sorts object arrays on a string key", () => {
-	const arr = [
-		{ title: "Quantitative theory" },
-		{ title: "Quantitative measurement" },
-		{ title: "Chaos theory" },
-		{ title: "Barometry" },
-		{ title: "Swarm intelligence" },
-		{ title: "Non-competitive signaling" }
-	];
-	expect(sortElems(arr, "title"))
-		.toEqual([
-			{ title: "Barometry" },
-			{ title: "Chaos theory" },
-			{ title: "Non-competitive signaling" },
-			{ title: "Quantitative measurement" },
-			{ title: "Quantitative theory" },
-			{ title: "Swarm intelligence" }
-		]);
 });
